G: SessionService; terminateSession should throw exception on invalid cookie

+++ app/src/main/java/expenses_tracker/app/session/SessionService.java
@@ -20,6 +20,9 @@ public class SessionService {
 
 	public Mono<Void> terminateSession(MultiValueMap<String, HttpCookie> cookies) {
 		HttpCookie sessionCookie = cookies.getFirst(sessionCookieName);
+		if (sessionCookie == null) {
+			return Mono.error(new IllegalArgumentException());
+		}
 		Mono<Void> result = repo.deleteSession(sessionCookie);
 		return result.onErrorResume(Exception.class,
 				e -> Mono.error(new IllegalArgumentException()));
